% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_agritox.R
\name{format_agritox}
\alias{format_agritox}
\title{Formatage du dataset AGRITOX en une table d'attributs par SA}
\usage{
format_agritox(x, source_name, modified_date)
}
\arguments{
\item{x}{dataframe jeu de données à formater}

\item{source_name}{char Nom de la source à afficher}

\item{modified_date}{date Nom de la source à afficher}
}
\value{
tibble
}
\description{
Formatage du dataset AGRITOX en une table d'attributs par SA
}
\examples{
library(magrittr)

# Lecture du dataset AGRITOX
agritox_path <- system.file("agritox/", package = "RefPesticidesBzh")
agritox <- RefPesticidesBzh::load_dataset(path=agritox_path, pattern="*.csv", encoding = "Latin-1")

# Ajout de la date de mise à jour à partir de la fiche Data.gouv
agritox$mise_a_jour <- RefPesticidesBzh::metadonnees_dataset('https://www.data.gouv.fr/fr/datasets/base-de-donnees-agritox/')$`Modifiée.le` \%>\% as.Date("\%d \%b \%Y")

# Mise en forme de la liste des SA avec les attributs retenus pour cette source
agritox_sa <- agritox \%>\% format_agritox(source_name = "AGRITOX", modified_date = agritox$mise_a_jour)

}
