% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Lecture d'un dataset importé}
\usage{
load_dataset(name, pattern = "*.csv")
}
\arguments{
\item{name}{char nom du dataset}

\item{pattern}{char motif de sélection des fichiers à importer}
}
\value{
Liste contenant les tables lues dans chaque fichier correspondant au motif
}
\description{
Lecture d'un dataset importé
}
\examples{

agritox <- load_dataset(name="agritox", pattern="*.csv")

ephy <- load_dataset(name="ephy", pattern="*_utf8.csv")

oeb_referentiels_substances_actives <- load_dataset(name="oeb_referentiels_substances_actives", pattern="*.csv")

# Adapter la fonction load_dataset au jeu de données Siris
siris <- readxl::read_xls("extdata/siris/siris.xls", skip=5)
names(siris) <- make.names(names(siris))

}
