% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substances_actives.R
\docType{data}
\name{substances_actives}
\alias{substances_actives}
\title{Liste de référence des substances actives pesticides}
\format{
Une table de données de 1140 lignes et 7 variables :
\describe{
\item{SA_CodeCAS}{Le numéro CAS (CAS number ou CAS registry number en anglais) d'une substance chimique est son numéro d'enregistrement unique auprès de la banque de données de Chemical Abstracts Service (CAS)}
\item{SA_Libelle}{Le nom chimique d’une substance active selon le standard IUPAC est le nom décrit en français d'après les règles les plus récentes de l'IUPAC (Union Internationale de Chimie Pure et Appliquée).}
\item{SA_PNEC}{}
\item{SA_Statut}{}
\item{source}{La source de données à afficher}
\item{mise_a_jour}{La date de mise à jour de la donnée}
}
}
\source{
\url{https://github.com/Observatoire-environnement-Bretagne/ReferentielPesticidesBzh}
}
\usage{
substances_actives
}
\description{
Table de description des substances actives retenues dans le cadre du groupe de travail régional animé par l'Observatoire de l'Environnement en Bretagne. Les différentes caractéristiques proposées sont issues d'une agrégation de différentes sources.
}
\keyword{datasets}
