% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{Fonction d'import d'un dataset à partir d'un fichier ZIP téléchargeable}
\usage{
import_dataset(url, name, format)
}
\arguments{
\item{url}{car url du jeu de données}

\item{name}{char nom du jeu de données}

\item{format}{format du fichier téléchargé (zip)}
}
\value{
Aucun
}
\description{
Fonction d'import d'un dataset à partir d'un fichier ZIP téléchargeable
}
\examples{
import_dataset('https://www.data.gouv.fr/fr/datasets/r/98dbe26f-a2e4-4002-a598-c226bf6af664','agritox','zip')
}
