% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_agritox.R, R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{Téléchargement et décompression du dataset depuis une url}
\usage{
import_dataset(url, name, format)

import_dataset(url, name, format)
}
\arguments{
\item{url}{char url stable du jeu de données}

\item{name}{char nom du dataset}

\item{format}{char format du fichier téléchargé (zip, xls, csv)}
}
\value{
Aucune reponse - Les fichiers sont téléchargés et dézippés dans le dossier extdata/

Aucune reponse - Les fichiers sont téléchargés et dézippés dans le dossier extdata/
}
\description{
Téléchargement et décompression du dataset depuis une url

Téléchargement et décompression du dataset depuis une url
}
\examples{
# Téléchargement du dataset agritox
import_dataset(url='https://www.data.gouv.fr/fr/datasets/r/ce799eee-2cc4-4406-b4b4-318ea35cd5e8', name="agritox", format="zip")

# Téléchargement du dataset ephy
import_dataset(url='https://www.data.gouv.fr/fr/datasets/r/98f7cac6-6b29-4859-8739-51b825196959', name="ephy", format="zip")

# Téléchargement du dataset "Description des substances actives phytosanitaires" de l'OEB
import_dataset(url='http://gide.bretagne-environnement.org/index.php/content/download/25084/384395/file/oeb_referentiels_substances_actives.csv', name="oeb", format="csv")

# Les fichiers sont téléchargés dans les sous-dossiers du répertoire de travail
# Téléchargement du dataset agritox
import_dataset(url='https://www.data.gouv.fr/fr/datasets/r/ce799eee-2cc4-4406-b4b4-318ea35cd5e8', name="agritox", format="zip")

# Téléchargement du dataset ephy
import_dataset(url='https://www.data.gouv.fr/fr/datasets/r/98f7cac6-6b29-4859-8739-51b825196959', name="ephy", format="zip")

# Téléchargement du dataset "Description des substances actives phytosanitaires" de l'OEB
import_dataset(url='http://gide.bretagne-environnement.org/index.php/content/download/25084/384395/file/oeb_referentiels_substances_actives.csv', name="oeb", format="csv")

# Les fichiers sont téléchargés dans les sous-dossiers du répertoire de travail
}
