% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_agritox.R, R/import_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Lecture d'un dataset importé}
\usage{
load_dataset(path, pattern = "*.csv", encoding = "UTF-8")

load_dataset(path, pattern = "*.csv", encoding = "UTF-8")
}
\arguments{
\item{path}{char chemin vers le répertoire des fichiers à charger}

\item{pattern}{char motif de sélection des fichiers}

\item{encoding}{char encodage des fichiers}
}
\value{
Liste contenant les tables lues dans chaque fichier correspondant au motif

Liste contenant les tables lues dans chaque fichier correspondant au motif
}
\description{
Lecture d'un dataset importé

Lecture d'un dataset importé
}
\examples{

agritox_path <- system.file("agritox/", package = "RefPesticidesBzh")
agritox_raw <- RefPesticidesBzh::load_dataset(path=agritox_path, pattern="*.csv", encoding = "UTF-8")


agritox_path <- system.file("agritox/", package = "RefPesticidesBzh")
agritox_raw <- RefPesticidesBzh::load_dataset(path=agritox_path, pattern="*.csv", encoding = "UTF-8")

}
