% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R, R/metadonnees_dataset.R
\name{metadonnees_dataset}
\alias{metadonnees_dataset}
\title{Metadonnées du dataset}
\usage{
metadonnees_dataset(url)

metadonnees_dataset(url)
}
\arguments{
\item{url}{char url de la fiche data.gouv.fr}
}
\value{
tibble des métadonnées du jeu de données

tibble des métadonnées du jeu de données
}
\description{
Metadonnées du dataset

Metadonnées du dataset
}
\examples{
metadonnees_dataset('https://www.data.gouv.fr/fr/datasets/base-de-donnees-agritox/')
metadonnees_dataset('https://www.data.gouv.fr/fr/datasets/base-de-donnees-agritox/')
}
